Refactor writing preview generation into separate lib functions
